Renamed 'hooks' to 'callbacks'.
