+++ classes/clean_moodle_url.php
@@ -108,13 +108,13 @@ class clean_moodle_url extends moodle_url {
         $this->path = $path;
     }
 
-    public static function find_format_hook($format) {
+    public static function find_format_callback($format) {
         $classname = "\\format_{$format}\\cleanurls_support";
         if (class_exists($classname)) {
             return $classname;
         }
 
-        $classname = "\\local_cleanurls\\local\\hooks\\{$format}_support";
+        $classname = "\\local_cleanurls\\local\\callbacks\\{$format}_support";
         if (class_exists($classname)) {
             return $classname;
         }
