+++ classes/cleaner.php
@@ -214,7 +214,7 @@ class cleaner {
         }
 
         // Try using a plugin hook (the plugin defines the behaviour) or a local hook.
-        $classname = clean_moodle_url::find_format_hook($course->format);
+        $classname = clean_moodle_url::find_format_callback($course->format);
         if (!is_null($classname)) {
             return '/' . $classname::get_clean_subpath($course, $cm);
         }
