+++ classes/uncleaner.php
@@ -212,7 +212,7 @@ class uncleaner {
     }
 
     private function unclean_course_format_hook(stdClass $course, array $parameters) {
-        $classname = clean_moodle_url::find_format_hook($course->format);
+        $classname = clean_moodle_url::find_format_callback($course->format);
         if (is_null($classname)) {
             return false;
         }
