+++ tests/local/hooks/format_flexsections_support_test.php
@@ -36,7 +36,7 @@ defined('MOODLE_INTERNAL') || die();
  * @copyright   2017 Catalyst IT Australia {@link http://www.catalyst-au.net}
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class local_cleanurls_hooks_flexsections_support extends advanced_testcase {
+class local_cleanurls_callbacks_flexsections_support extends advanced_testcase {
     protected function setUp() {
         global $CFG;
         parent::setUp();
